Create 06_TEST_QUESTION.sv
Implement a code where two parallel threads are running and one thread is sending data while the other thread is receiving it. Assume the data to be a structure which consists of information 32-bit data and 8-bit lenght. The mechanism should work in such a way that, at any point of time the sender can send only send the next structure/pkt only when the previous one has been received by the receiver Assume that the receiver will accept the pkt after some random amount of delay for every pkt.